VPC: use a different key for the network
There was no version check before, so an older release of sevenseconds
would happily destroy accounts that have custom subnet setups. Instead
of hoping that all the users update, change the configuration format
slightly so old versions would throw an exception instead. This will not
be necessary in the future because we now have an explicit check at the
beginning.

Signed-off-by: Alexey Ermakov <alexey.ermakov@zalando.de>