Add option to disable odd per account
With more accounts in Kubernetes and other soltions like EC2 Instance
Connect odd hosts are not needed. This commit creates the `enable_odd`
config that defaults to `True` and enable us to disable this feature per
account. If odd instances are found and it's disabled it will be
removed while in new accounts it won't be created.

Signed-off-by: Jonathan Juares Beber <jonathanbeber@gmail.com>